+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "my-app",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -16,6 +16,8 @@
     "@react-navigation/native-stack": "^7.3.3",
     "@react-navigation/stack": "^7.2.8",
     "@shadcn/ui": "^0.0.4",
+    "@types/axios": "^0.14.4",
+    "axios": "^1.8.4",
     "expo": "~52.0.42",
     "expo-camera": "~16.0.18",
     "expo-file-system": "^18.0.12",
