+++ src/App.tsx
@@ -2,8 +2,9 @@ import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { StyleSheet, StatusBar } from "react-native";
-import { UserProvider } from "../my-app/screens/user";
-import { RootNavigator } from "../my-app/navigation/AppNavigation"
+import { UserProvider } from "./src/screens/user";
+import { RootNavigator } from "./src/navigation/AppNavigation";
+
 
 export function App() {
   return (
