+++ src/screens/auth/.index.ts
@@ -1,0 +1,4 @@
+export * from "./EmailScreen"
+export * from "./Login"
+export * from "./VerificationScreen"
+export * from "./singupScreen"
