+++ src/screens/banco.tsx
