+++ src/types/interfaces/UserPostInterface.ts
@@ -1,0 +1,5 @@
+export interface EmailPostInterface {
+    Email:string,
+    Password:string,
+    userName:string
+}
