+++ src/types/rootStackParamList.ts
@@ -1,0 +1,8 @@
+export type RootStackParamList = {
+  Welcome: undefined;
+  Login: undefined;
+  Signup: undefined;
+  Recuperação: undefined
+  Home: undefined;
+  Email: undefined;
+};
